+++ src/Router.js
@@ -2,15 +2,14 @@ import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Main from './pages/Main/Main';
 import Login from './pages/Login/Login';
-import SignUp from './pages/SignUp/SignUp';
 
 const Router = () => {
   return (
     <BrowserRouter>
       <Routes>
         <Route path="/login" element={<Login />} />
+        <Route path="/signup" element={<Login />} />
         <Route path="/main" element={<Main />} />
-        <Route path="/signup" element={<SignUp />} />
       </Routes>
     </BrowserRouter>
   );
