+++ src/pages/Login/Login.scss
@@ -34,12 +34,12 @@
 
   input,
   button,
-  .loginText {
+  .title {
     width: 375px;
     height: 40px;
   }
 
-  .loginText {
+  .title {
     font-size: 30px;
     font-weight: 700;
     line-height: 1.33;
