--- src/pages/SignUp/SignUp.js
@@ -1,7 +1,0 @@
-import React from 'react';
-import './SignUp.scss';
-function SignUp() {
-  return <div>SignUp</div>;
-}
-
-export default SignUp;
